clean codex

+++ frameworks/proxy/event_handler/src/input_manager_impl.cpp
@@ -170,12 +170,12 @@ int32_t InputManagerImpl::PackDisplayData(OHOS::MMI::NetPacket &pkt)
 
 int32_t InputManagerImpl::PackPhysicalDisplay(NetPacket &pkt)
 {
-    int32_t num = physicalDisplays_.size();
+    uint32_t num = static_cast<uint32_t>(physicalDisplays_.size());
     if (!pkt.Write(num)) {
         MMI_LOGE("Packet write num failed");
         return RET_ERR;
     }
-    for (int32_t i = 0; i < num; i++) {
+    for (uint32_t i = 0; i < num; i++) {
         if (!pkt.Write(physicalDisplays_[i].id)) {
             MMI_LOGE("Packet write physical data failed");
             return RET_ERR;
